Move test closer to code under test
