+++ crates/fj-core/src/validate/shell.rs
@@ -272,28 +272,6 @@ mod tests {
     };
 
     #[test]
-    fn half_edge_has_no_sibling() -> anyhow::Result<()> {
-        let mut core = Core::new();
-
-        let valid = Shell::tetrahedron(
-            [[0., 0., 0.], [0., 1., 0.], [1., 0., 0.], [0., 0., 1.]],
-            &mut core,
-        );
-        let invalid = valid.shell.remove_face(&valid.abc.face);
-
-        valid
-            .shell
-            .validate_and_return_first_error(&core.layers.geometry)?;
-        assert_contains_err!(
-            core,
-            invalid,
-            ValidationError::HalfEdgeHasNoSibling { .. }
-        );
-
-        Ok(())
-    }
-
-    #[test]
     fn coincident_half_edges_are_not_siblings() -> anyhow::Result<()> {
         let mut core = Core::new();
